### 2026-02-18 426fb91
fix: rOpenSci pkgcheck compliance — complexity, coverage, governance
Address 3 blocking pkgcheck issues for rOpenSci submission:

- Add CONTRIBUTING.md and CODE_OF_CONDUCT.md governance files
- Reduce cyclomatic complexity of all 11 high-complexity exported
  functions (sipni_data 60→6, vigitel_data 33→13, pnadc_data 27→8,
  pns_data 23→5, pof_data 23→10, sisab_data 23→2, sim_data 21→7,
  sia_data 17→11, sih_data 15→13, pns_dictionary 19→7,
  pns_sidra_data 15→9) by extracting internal subfunctions
- Increase test coverage from 54% to 82% with ~1300 new tests
  across all modules (3286 total)

[…]
