### 2026-02-18 dc75ca9
refactor: reduce all internal functions to cyclomatic complexity < 15
Extract subfunctions from 10 internal helpers that exceeded the
threshold:
- sipni.R: .sipni_csv_process_national (34→5),
  .sipni_csv_download_months (23→3), .sipni_try_lazy_pre (17→12),
  .sipni_download_and_read (16→10)
- pof.R: .pof_create_survey_design (25→10),
  .pof_find_dictionary_file (25→5), .pof_parse_dictionary (23→14)
- utils-download.R: .http_download_resumable (28→3),
  .datasus_download (15→1), .http_download (15→1) via shared
  .retry_download_loop() helper

goodpractice::gp(checks = "cyclocomp") now passes 
[…]
