+++ R/pof.R
@@ -228,24 +228,27 @@ pof_cache_dir <- function(cache_dir = NULL) {
   zip_file
 }
 
-#' Find dictionary file in extracted documentation
+#' Find dictionary file by direct grep pattern matching
+#' @param xls_files Character vector of .xls file paths.
+#' @param patterns Character vector of patterns to try.
+#' @return Matched file path or NULL.
 #' @noRd
-.pof_find_dictionary_file <- function(extracted_files, year) {
-  # look for dictionary Excel file
-  # use flexible patterns to handle encoding issues with Portuguese characters
-  xls_files <- extracted_files[grepl("\\.xls", extracted_files, ignore.case = TRUE, useBytes = TRUE)]
-
-  # approach 1: direct pattern matching
-  dict_patterns <- c("dicion", "variav", "variaveis")
-  for (pattern in dict_patterns) {
+.pof_match_xls_by_pattern <- function(xls_files, patterns) {
+  for (pattern in patterns) {
     for (f in xls_files) {
       if (grepl(pattern, f, ignore.case = TRUE, useBytes = TRUE)) {
         return(f)
       }
     }
   }
+  NULL
+}
 
-  # approach 2: check raw bytes for "Dicion" (handles encoding issues)
+#' Find dictionary file by raw byte comparison for "Dicion"
+#' @param xls_files Character vector of .xls file paths.
+#' @return Matched file path or NULL.
+#' @noRd
+.pof_match_xls_by_bytes <- function(xls_files) {
   dicion_raw <- charToRaw("Dicion")
   for (f in xls_files) {
     bn <- basename(f)
@@ -256,15 +259,99 @@ pof_cache_dir <- function(cache_dir = NULL) {
       }
     }
   }
+  NULL
+}
 
-  # approach 3: look for files with "variav" in accent-stripped name
+#' Find dictionary file by accent-stripped name matching
+#' @param xls_files Character vector of .xls file paths.
+#' @return Matched file path or NULL.
+#' @noRd
+.pof_match_xls_by_stripped <- function(xls_files) {
   for (f in xls_files) {
     bn_ascii <- .strip_accents(basename(f))
     if (grepl("dicion|variav", bn_ascii, ignore.case = TRUE, useBytes = TRUE)) {
       return(f)
     }
   }
+  NULL
+}
 
+#' Find dictionary file in extracted documentation
+#' @noRd
+.pof_find_dictionary_file <- function(extracted_files, year) {
+  # look for dictionary Excel file
+  # use flexible patterns to handle encoding issues with Portuguese characters
+  xls_files <- extracted_files[grepl("\\.xls", extracted_files, ignore.case = TRUE, useBytes = TRUE)]
+
+  # approach 1: direct pattern matching
+  result <- .pof_match_xls_by_pattern(xls_files, c("dicion", "variav", "variaveis"))
+  if (!is.null(result)) return(result)
+
+  # approach 2: check raw bytes for "Dicion" (handles encoding issues)
+  result <- .pof_match_xls_by_bytes(xls_files)
+  if (!is.null(result)) return(result)
+
+  # approach 3: look for files with "dicion" or "variav" in accent-stripped name
+  .pof_match_xls_by_stripped(xls_files)
+}
+
+#' Standardize dictionary column names to canonical format
+#'
+#' Renames columns from the raw Excel dictionary to standard names:
+#' position, length, variable, description, decimals, categories.
+#' @param df Data frame with raw column names from readxl.
+#' @return The same data frame with standardized column names.
+#' @noRd
+.pof_standardize_dict_columns <- function(df) {
+  col_names <- names(df)
+
+  # mapping: standard_name -> grep pattern
+  rename_map <- list(
[…]
