+++ R/sipni.R
@@ -99,6 +99,28 @@
 }
 
 
+#' Check FTP partitioned cache for SI-PNI data
+#' @return Cached tibble or NULL if not found.
+#' @noRd
+.sipni_ftp_check_cache <- function(uf, year, type, cache, cache_dir) {
+  dataset_name <- stringr::str_c("sipni_", tolower(type), "_data")
+  target_year <- as.integer(year)
+
+  if (!isTRUE(cache) || !.has_arrow() ||
+      !.has_partitioned_cache(cache_dir, dataset_name)) {
+    return(NULL)
+  }
+
+  ds <- arrow::open_dataset(file.path(cache_dir, dataset_name))
+  cached <- ds |>
+    dplyr::filter(.data$uf_source == uf,
+                   .data$year == target_year) |>
+    dplyr::collect()
+
+  if (nrow(cached) > 0) cached else NULL
+}
+
+
 #' Download and read a SI-PNI .DBF file for one UF/year/type (FTP, 1994-2019)
 #' @noRd
 .sipni_download_and_read <- function(year, uf, type = "DPNI",
@@ -109,15 +131,8 @@
   target_uf <- uf
 
   # 1. check partitioned cache first (preferred path)
-  if (isTRUE(cache) && .has_arrow() &&
-      .has_partitioned_cache(cache_dir, dataset_name)) {
-    ds <- arrow::open_dataset(file.path(cache_dir, dataset_name))
-    cached <- ds |>
-      dplyr::filter(.data$uf_source == target_uf,
-                     .data$year == target_year) |>
-      dplyr::collect()
-    if (nrow(cached) > 0) return(cached)
-  }
+  cached <- .sipni_ftp_check_cache(uf, year, type, cache, cache_dir)
+  if (!is.null(cached)) return(cached)
 
   # 2. download from FTP
   url <- .sipni_build_ftp_url(year, uf, type)
@@ -188,46 +203,37 @@
 }
 
 
-#' Process national SI-PNI CSV for one month, caching ALL UFs
-#'
-#' Downloads the national monthly CSV ZIP from OpenDataSUS, reads it in
-#' chunks splitting by UF, and caches ALL 27 states. This means a second
-#' request for a different UF from the same month is instant from cache.
-#'
-#' @param year Integer. Year.
-#' @param month Integer. Month (1-12).
-#' @param uf Character. The UF to return (all 27 are cached regardless).
-#' @param cache Logical. Whether to use caching.
-#' @param cache_dir Character or NULL. Cache directory.
-#' @param zip_path Character or NULL. Path to an already-downloaded ZIP file.
-#'   If provided, skips downloading. Used by `.sipni_csv_download_months()`
-#'   for concurrent download + sequential processing.
-#'
-#' @return A tibble with data for the requested UF.
+#' Check partitioned cache for SI-PNI CSV data
+#' @return Cached tibble or NULL if not found.
 #' @noRd
-.sipni_csv_process_national <- function(year, month, uf,
-                                        cache = TRUE, cache_dir = NULL,
-                                        zip_path = NULL) {
-  cache_dir <- .sipni_cache_dir(cache_dir)
+.sipni_csv_check_cache <- function(uf, year, month, cache, cache_dir) {
   dataset_name <- "sipni_csv_data"
   target_year <- as.integer(year)
   target_month <- as.integer(month)
-  month_name <- sipni_month_names[month]
 
-  # 1. check partitioned cache first (preferred path)
-  if (isTRUE(cache) && .has_arrow() &&
-      .has_partitioned_cache(cache_dir, dataset_name)) {
-    ds <- arrow::open_dataset(file.path(cache_dir, dataset_name))
-    cached <- ds |>
-      dplyr::filter(.data$uf_source == uf,
-                     .data$year == target_year,
[…]
