+++ pkgdown.yml
@@ -19,7 +19,7 @@ articles:
   sipni-vaccination: sipni-vaccination.html
   sisab-primary-care: sisab-primary-care.html
   vigitel-telephone-survey: vigitel-telephone-survey.html
-last_built: 2026-02-18T14:09Z
+last_built: 2026-02-18T14:31Z
 urls:
   reference: https://sidneybissoli.github.io/healthbR/reference
   article: https://sidneybissoli.github.io/healthbR/articles
