### 2026-02-18 e9f2a05
fix: replace setwd() with withr::with_dir() in PNS tests
Addresses goodpractice lint warning about global state mutation.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
