### 2026-02-18 b95766b
fix: guard file.size() with file.exists() and use writeBin in SI-PNI test mocks
Fixes 3 test failures on Ubuntu CI where file.copy() in mocked
.datasus_download silently failed, leaving temp_dbf non-existent and
causing file.size() to return NA. Tests now capture DBF as raw bytes
upfront and use writeBin() for reliable cross-platform behavior.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
