### 2026-02-18 0bc04d5
fix: use pre-computed DBF bytes in SI-PNI tests for Linux CI compatibility
foreign::write.dbf() silently fails on Ubuntu CI runners, causing
test DBF files to not be created. Replace with base64-encoded raw DBF
bytes decoded via jsonlite::base64_dec(), eliminating all filesystem
dependencies in the test setup.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
