### 2026-02-18 e1b084d
docs: update README with all 16 modules and update NEWS changelog
README: add SINAN, CNES, SI-PNI to DATASUS table; add SISAB, ANS,
ANVISA section; add quick-start examples for all new modules; update
data sources list and governance links.

NEWS: add cyclomatic complexity refactoring, governance files, and
SI-PNI file.exists() bug fix entries.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
