+++ articles/healthbR.md
@@ -7,7 +7,7 @@ from R. The package downloads, caches, and processes data from official
 sources (Ministry of Health, IBGE, DATASUS), returning clean,
 analysis-ready tibbles following tidyverse conventions.
 
-The package currently supports **13 data sources** organized in two
+The package currently supports **16 data sources** organized in three
 groups:
 
 **Surveys (IBGE / Ministry of Health)**
@@ -33,6 +33,13 @@ groups:
 | SI-PNI | Vaccination (aggregated FTP + microdata API) | Annual/UF       | 1994–2025    |
 | SISAB  | Primary care coverage indicators (REST API)  | Monthly         | 2007–present |
 
+**Regulatory agencies**
+
+| Module | Source                                                      | Granularity | Years        |
+|--------|-------------------------------------------------------------|-------------|--------------|
+| ANS    | Supplementary health (beneficiaries, complaints, financial) | Monthly/UF  | 2007–present |
+| ANVISA | Medicines, devices, pharmacovigilance, SNGPC                | National    | 2014–present |
+
 ## Getting started
 
 ``` r
@@ -44,7 +51,7 @@ To see all available data sources at a glance:
 
 ``` r
 list_sources()
-#> # A tibble: 13 x 5
+#> # A tibble: 16 x 5
 #>    source  name                         description                    years       status
 #>    <chr>   <chr>                        <chr>                          <chr>       <chr>
 #>  1 vigitel VIGITEL                      Telephone survey on chronic... 2006-2024   available
@@ -357,6 +364,62 @@ sisab_variables()
 sisab_variables(type = "sb")
 ```
 
+## ANS – Supplementary health
+
+ANS (Agencia Nacional de Saude Suplementar) provides open data on
+private health insurance in Brazil, including beneficiary counts,
+consumer complaints, and operator financial statements.
+
+``` r
+# module overview
+ans_info()
+
+# beneficiaries by state, January 2024
+benef <- ans_data(type = "beneficiaries", year = 2024, month = 1, uf = "SP")
+
+# consumer complaints, 2023
+reclamacoes <- ans_data(type = "complaints", year = 2023)
+
+# operator financial data, 2024 Q1
+financeiro <- ans_data(type = "financial", year = 2024, quarter = 1)
+
+# health plan operator registry
+operadoras <- ans_operators()
+
+# explore variables
+ans_variables(type = "beneficiaries")
+```
+
+See
+[`vignette("ans-supplementary-health")`](https://sidneybissoli.github.io/healthbR/articles/ans-supplementary-health.md)
+for detailed examples.
+
+## ANVISA – Medicines and pharmacovigilance
+
+ANVISA (Agencia Nacional de Vigilancia Sanitaria) publishes open data on
+regulated products (medicines, medical devices, food, cosmetics) and
+pharmacovigilance reports (VigiMed). The SNGPC module provides
+controlled substance dispensing data.
+
+``` r
+# see all available data types
+anvisa_types()
+
+# registered medicines
+medicamentos <- anvisa_data(type = "medicines")
+
+# pharmacovigilance notifications (VigiMed)
+vigimed <- anvisa_data(type = "vigimed_notifications")
+
+# controlled substance dispensing (SNGPC), monthly time-series
+sngpc <- anvisa_data(type = "sngpc", year = 2024, month = 1:6)
+
+# explore variables
+anvisa_variables(type = "medicines")
+```
+
+See `vignette("anvisa-regulatory-data")` for detailed examples.
+
 ## Caching
 
 All modules cache downloaded data automatically to avoid repeated
@@ -379,6 +442,47 @@ sim_clear_cache()
[…]
