### 2026-02-21 5b5e34c
docs: document parallel downloads in all *_data() help pages and README
Add "Parallel downloads" section to roxygen @details of all 12 *_data()
functions and to README.md, making the furrr/future opt-in mechanism
discoverable from ?function help and the package landing page.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
