+++ R/cnes.R
@@ -374,6 +374,12 @@ cnes_dictionary <- function(variable = NULL) {
 #' The CNES has 13 file types. The default \code{"ST"} (establishments) is
 #' the most commonly used. Use \code{\link{cnes_info}()} to see all types.
 #'
+#' ## Parallel downloads
+#' When downloading multiple files (e.g., several months or states), install
+#' \pkg{furrr} and \pkg{future} and set a parallel plan to speed up downloads:
+#' `future::plan(future::multisession, workers = 4)`. See
+#' `vignette("healthbR")` for details.
+#'
 #' @export
 #' @family cnes
 #'
