+++ reference/pnadc_data.md
@@ -105,6 +105,14 @@ following variables are always included:
 Use `as_survey = TRUE` to get a properly weighted survey design object
 for analysis with the `srvyr` package.
 
+### Parallel downloads
+
+When downloading multiple years, install furrr and future and set a
+parallel plan to speed up downloads:
+`future::plan(future::multisession, workers = 4)`. See
+[`vignette("healthbR")`](https://sidneybissoli.github.io/healthbR/articles/healthbR.md)
+for details.
+
 ## Data source
 
 Data is downloaded from the IBGE FTP server:
