+++ reference/cnes_data.md
@@ -114,6 +114,14 @@ most commonly used. Use
 [`cnes_info()`](https://sidneybissoli.github.io/healthbR/reference/cnes_info.md)
 to see all types.
 
+### Parallel downloads
+
+When downloading multiple files (e.g., several months or states),
+install furrr and future and set a parallel plan to speed up downloads:
+`future::plan(future::multisession, workers = 4)`. See
+[`vignette("healthbR")`](https://sidneybissoli.github.io/healthbR/articles/healthbR.md)
+for details.
+
 ## See also
 
 [`cnes_info()`](https://sidneybissoli.github.io/healthbR/reference/cnes_info.md)
