+++ reference/sipni_data.md
@@ -122,6 +122,14 @@ vaccination dose, ~47 fields per record). The `type` parameter is
 ignored for CSV years. Data is filtered by UF during chunked reading to
 avoid loading the full national file into memory.
 
+### Parallel downloads
+
+When downloading multiple files (e.g., several years or states), install
+furrr and future and set a parallel plan to speed up downloads:
+`future::plan(future::multisession, workers = 4)`. See
+[`vignette("healthbR")`](https://sidneybissoli.github.io/healthbR/articles/healthbR.md)
+for details.
+
 ## See also
 
 [`sipni_info()`](https://sidneybissoli.github.io/healthbR/reference/sipni_info.md)
