### 2026-02-21 a11e97a
feat: add download progress bars to all modules via .map_parallel()
Shows spinner, current/total count, progress bar, and ETA during
multi-file downloads using cli::cli_progress_bar(). Covers all 16
modules (14 call sites). Progress is skipped for single-file downloads.

Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
