+++ R/sih.R
@@ -307,7 +307,8 @@ sih_dictionary <- function(variable = NULL) {
   labels <- paste(combinations$uf,
                   paste0(combinations$year, "/", sprintf("%02d", combinations$month)))
 
-  results <- .map_parallel(seq_len(n_combos), .delay = 0.5, function(i) {
+  results <- .map_parallel(seq_len(n_combos), .delay = 0.5,
+                            .progress = "Downloading", function(i) {
     yr <- combinations$year[i]
     mo <- combinations$month[i]
     st <- combinations$uf[i]
