+++ R/sinasc.R
@@ -396,7 +396,8 @@ sinasc_data <- function(year, vars = NULL, uf = NULL, anomaly = NULL,
   # download and read each combination
   labels <- paste(combinations$uf, combinations$year)
 
-  results <- .map_parallel(seq_len(n_combos), .delay = 0.5, function(i) {
+  results <- .map_parallel(seq_len(n_combos), .delay = 0.5,
+                            .progress = "Downloading", function(i) {
     yr <- combinations$year[i]
     st <- combinations$uf[i]
 
