+++ R/cnes.R
@@ -442,7 +442,8 @@ cnes_data <- function(year, type = "ST", month = NULL, vars = NULL, uf = NULL,
   labels <- paste(type, combinations$uf,
                   paste0(combinations$year, "/", sprintf("%02d", combinations$month)))
 
-  results <- .map_parallel(seq_len(n_combos), .delay = 0.5, function(i) {
+  results <- .map_parallel(seq_len(n_combos), .delay = 0.5,
+                            .progress = "Downloading", function(i) {
     yr <- combinations$year[i]
     mo <- combinations$month[i]
     st <- combinations$uf[i]
