### 2026-02-21 #751 @SidneyBissoli
Hi, Emily.

Thanks for getting in touch. I was eager to hear back. Here are my answers:
1. I totally agree.
2. I think this question requires a longer answer, so let's go:

As you can see, Claude was used extensively as an implementation tool
throughout development, which is why it appears as co-author in commits.

What I have done:
 - Package conception, scope, and module selection (which 16 Brazilian
health data sources to support)
 - API design decisions (function signatures, parameter naming, return
types)
 - Architectural choices (Hive-style partitioned parquet caching, lazy
evaluation via Arrow/DuckDB, DBC decompression strategy)
 - Domain knowledge of Brazilian health data systems (DATASUS file naming
conventions, FTP directory structures, encoding quirks, variable
dictionaries)
 - Code review and acceptance/rejection of all generated code
 - Testing against real data sources and validation of outputs
 - Bug discovery and prioritization (e.g., identifying Windows encoding
issues, FTP edge cases)

 What Claude has done:
 - Writing function bodies and boilerplate from my specifications
 - Writing unit and integration tests from my test plans
 - Debugging (e.g., diagnosing iconv segfaults on Windows, Arrow
partitioning issues)
 - Refactoring tasks (e.g., extracting shared validation/search/cache
helpers across modules)
 - Documentation drafting (roxygen2 help pages, vignettes)

In practice, the workflow was iterative: I would describe what a module
should do, how the data source works, and what the function interface
should look like. Claude would produce an implementation. I would test it
against real data, identify issues, and direct fixes. All code was reviewed
by me before being committed, and the "Co-Authored-By" tag reflects this
collaborative process.

Your last question was: why generative AI tools were used? Short and direct
answer: because I wouldn't be able to do all of this work alone. I started
learning R in 2020, during the pandemic. But I have never had an experience
of building a package. And getting Brazilian public health data is a real
and big pain (at least, it was for me, specially during my R learning
path). Now, I am taking a course about Building R Packages, at Coursera.
For me, submitting this R package to rOpenSci can serve as a learning
experience.

My best regards,

Sidney

Em sex., 20 de fev. de 2026 às 21:55, Emily Riederer <
***@***.***> escreveu:

> *emilyriederer* left a comment (ropensci/software-review#751)
> <https://github.com/ropensci/software-review/issues/751#issuecomment-3937816713>
>
> Hi @SidneyBissoli <https://github.com/SidneyBissoli> ! Thank you for your
> submission.
>
> Overall, I think your package meets our specifications for review.
> However, I have a few initial inquiries.
>
>    1. Could we limit the scope to "data retrieval"? Per our definitions
>    <https://devguide.ropensci.org/softwarereview_policies.html#aims-and-scope>,
>    retrieval refers to web data such as APIs whereas extraction is moreso
>    structuring data from unstructured files.
>    2. We are also currently in the process of defining our approach to
>    reviews for packages developed with the use of AI. Although the guidance is
>    not formally published, we do plan to require authors to explicitly
>    describe how, where, and why generative AI tools were used. (This is not a
>    value judgement on using AI but instructive to think about how to approach
>    the review). I noticed that Claude is attributed a substantial proportion
>    of the commits in your repo. Could you please provide a high-level
>    explanation to how you distinguished human-generated and AI-generated tasks
>    while developing your package?
>
