### 2026-02-24 #751 @SidneyBissoli
Hi, Emily!

Not in days, but surely not in months (or, at least, just a few months),
and I am trying to solve it right now. Let me explain in more detail to
you.

The problem is: it is incredibly slow how my R package takes to get .dbf
files (format from Brazilian Health's Ministry) from FTP, and stores them
in local cache as partitioned Parquet files (by year, month and States). I
have thought this way for one reason: subsequent reads use Apache Arrow
with lazy evaluation, allowing pushdown filters that load only the
requested subsets, without needing to read the entire dataset into memory.
But the first call was painful. And just for one year of immunization
microdata.

So, I am going to change the whole architecture, and divide this workflow
into three steps:

Step 1: Convert csv files from Brazilian Health's Ministry (in this case,
not from FTP) to partitioned Parquet format (again, by year, month and
States) and store them in R2 (Cloudfare). I will pay a VPS to upload data
(in fact, I have already paid). The pipeline already predicts one month of
data actualization (same frequency as Brazilian Health's Ministry),
checking what is new file, what is the same file (in name and size) and
what is just apparently the same (same name but different size). This is a
way to ensure that all stored data will be completely identical to that of
the Ministry of Health.

Step 2: I will also provide this storage with aggregated data (for
instance: countings by year, month, States, etc.).

Steps 1 and 2 will be properly documented in a GitHub repository (not even
the pipeline but also official documents from Ministry of Health and data
dictionary, README files, etc.), it is already being made with R code, but
they will not comprise the healthbR package itself (it is just about
showing the project, documenting it for transparency, not about sharing R
functions in a package).

Step 3: R researchers (or anyone) will use healthbR to get health data
directly from Parquet files stored in R2 (Cloudfare). If she/he wants
aggregated data, she/he will get it, immediately (for example: no tons of
work to get vaccination coverage). But if she/he wants microdata, she/he
will also get, in a much faster way than obtaining it from Ministry of
Health (or from the healthbR package as it is now).

Unfortunately, I have also published the package on CRAN, so I am trying to
change its infrastructure as soon as I can. It will be costly for me, not
only in terms of time availability, but also financially. But I accept this
challenge.

Sorry for the enormous e-mail and for taking your time.

My best wishes,

Sidney

Em seg., 23 de fev. de 2026 às 20:21, Emily Riederer <
***@***.***> escreveu:

> *emilyriederer* left a comment (ropensci/software-review#751)
> <https://github.com/ropensci/software-review/issues/751#issuecomment-3947889255>
>
> Hey @SidneyBissoli <https://github.com/SidneyBissoli> ! Thanks for the
> incredibly thoughtful and detailed response re: AI and for letting me know
> about the issues. That sounds like a great next step. What sort of a
> timeframe are you thinking to resubmit? If we are talking days-weeks, I can
> leave this issue open. Otherwise, I might close it for now for tracking
> purposes and allow us to start fresh once you resubmit.
>
> FWIW, part of the reason for my questions above is that we are
> simultaneously developing our strategies for reviewing packages with
> LLM-assisted code generation. Happily, we will probably have more thoughts
> here by the time you circle back!
>
