### 2026-02-24 #751 @emilyriederer
Hi @SidneyBissoli -- Good for you! That sounds like quite a large undertaking, but it's great that you are tackling major pain points like speed for your users. It's very interesting to hear how you are thinking about speed while preserving the ability for users to always go back to the true source and validate the two if they want. I don't often see authors consider the trade-off between building an API wrapper / data retrieving code versus a data product, which is a great reminder that open tools can exist at all levels of the stack. I also appreciate you're making this big infrastructure decision before beginning the review process so we can ensure the version reviewed is most useful to you.

Tactically, I'll leave this open but apply the "holding" tag just for tracking purposes. Please let us know when you're ready to pick up the discussion again.
