### 2026-05-03 #751 @maurolepore
Just touching base on the hold from February. No rush — please reach out whenever you're ready to proceed.

For your information, our holding policy (so the 1-year mark doesn't come as a surprise later):

> The author can choose to have their submission put on hold (editor applies the holding label). The holding status will be revisited every 3 months, and after one year the issue will be closed.
> — https://devguide.ropensci.org/policies#policiesreviewprocess

Thanks for sharing your work with the rOpenSci community.
