### 2026-07-31 e51334e
docs: adiciona CLAUDE.md e guia de integracao de dados
Co-Authored-By: Claude Fable 5 <noreply@anthropic.com>
