+++ inst/healthbr-data-integration.md
@@ -1,0 +1,242 @@
+# Integração healthbR ↔ healthbr-data (R2)
+
+> Documento de planejamento. Descreve a arquitetura proposta para integrar
+> o backend Cloudflare R2 do projeto healthbr-data ao pacote healthbR,
+> permitindo que o usuário escolha a fonte dos dados via parâmetro `source`.
+>
+> Status: **planejado** — nenhuma linha de código foi alterada ainda.
+> Criado em: 2026-03-07.
+
+---
+
+## 1. CONTEXTO
+
+O pacote `healthbR` acessa dados do SI-PNI de duas formas atualmente:
+
+- **FTP DATASUS (1994–2019):** arquivos `.dbf` baixados sob demanda por
+  UF × ano.
+- **OpenDATASUS CSV (2020+):** arquivos CSV mensais nacionais (~1.4 GB cada),
+  lidos em chunks e filtrados por UF.
+
+O projeto **healthbr-data** mantém os mesmos dados pré-processados e
+publicados como Parquet no Cloudflare R2 (`healthbr-data` bucket), com
+as seguintes vantagens:
+
+| Propriedade | Backend atual (FTP/CSV) | Backend R2 (Parquet) |
+|-------------|:-----------------------:|:--------------------:|
+| Microdados 2020+ | CSV com artefatos (`.0`, zeros perdidos em 2020–2024) | JSON origin — sem artefatos |
+| Agregados 1994–2019 | `.dbf` via FTP (lento, instável em BAs/MG/SP) | Parquet prontos — leitura instantânea |
+| Acesso offline | Não | Sim (após cache local) |
+| Leitura parcial (Arrow lazy) | Limitada | Nativa via `arrow::open_dataset()` |
+| Volume para múltiplos anos/UFs | Download sequencial | Scan colunar filtrado |
+
+---
+
+## 2. DESIGN PROPOSTO
+
+### 2.1 Mudança na assinatura de `sipni_data()`
+
+Adicionar dois novos parâmetros ao fim da assinatura existente:
+
+```r
+sipni_data(
+  year, type = "DPNI", uf = NULL, month = NULL,
+  vars = NULL, parse = TRUE, col_types = NULL,
+  cache = TRUE, cache_dir = NULL,
+  lazy = FALSE, backend = c("arrow", "duckdb"),
+  source = c("datasus", "r2"),   # NOVO
+  r2_credentials = NULL          # NOVO
+)
+```
+
+**`source`**  
+- `"datasus"` (padrão): comportamento atual, sem quebra de compatibilidade.  
+- `"r2"`: lê do Cloudflare R2 via Arrow S3 filesystem.
+
+**`r2_credentials`**  
+- Lista com `access_key_id` e `secret_access_key`.  
+- Se `NULL`, usa o token read-only público do `healthbr-data` como padrão
+  (o token é intencionalmente publicado; só permite leitura do bucket).  
+- Permite que usuários com credenciais próprias apontem para buckets
+  alternativos.
+
+### 2.2 Comportamento por `source`
+
+| `source` | Dados 1994–2019 | Dados 2020+ | Dicionários |
+|----------|-----------------|-------------|-------------|
+| `"datasus"` | `.dbf` via FTP | CSV via OpenDATASUS | internos (`sipni_data_internal.R`) |
+| `"r2"` | `sipni/agregados/doses/` ou `sipni/agregados/cobertura/` | `sipni/microdados/` | `sipni/dicionarios/` |
+
+### 2.3 Mapeamento de parâmetros existentes para o R2
+
+| Parâmetro `sipni_data()` | Comportamento no R2 |
+|--------------------------|---------------------|
[…]
