### 2026-08-19 #751 @SidneyBissoli
Thanks for the summary and the guidance, @maurolepore — closing this and starting from scratch later works for me.

A brief status note for the record: development continues. healthbR [0.3.0 is out](https://github.com/SidneyBissoli/healthbR/releases/tag/v0.3.0). The Ministry of Health decommissioned the OpenDataSUS host and removed the 2020–2025 immunization microdata from the new location, so the SI-PNI module now reads by default from [healthbr-data](https://github.com/SidneyBissoli/healthbr-data) — a companion data project that redistributes DATASUS/OpenDataSUS datasets as hive-partitioned Parquet on Cloudflare R2, byte-identical to the Ministry's files, with per-file provenance metadata and public manifests — falling back to the official sources automatically.

The plan is to extend this backend to the mortality (SIM), live births (SINASC) and hospital admissions (SIH) modules, and to restart the submission once that architecture has settled. I'll ping you then.
